🐛 Fix: MSVC compile bug

+++ lib/common/test/file.cpp
@@ -12,7 +12,7 @@ TEST_CASE("Write and read")
 	/* Generate Random Data */
 
 	std::mt19937_64 rng(114514);
-	std::uniform_int_distribution<uint8_t> dist(0, std::numeric_limits<uint8_t>::max());
+	std::uniform_int_distribution<uint16_t> dist(0, std::numeric_limits<uint8_t>::max());
 
 	std::vector<uint8_t> random_data;
 	random_data.reserve(8192);
